Added base functionality, mostly described in the README,
When running, will prompt for a YouTube video link, and for how many n most common words you want to see.
If the YT link is in the wrong format, will print an error message.

Otherwise, if all is good, will make an API call in analyzer.run(),
			parses the video transcript in analyzer.parse(),
			and prints it in analyzer.output()

YT links can come in multiple formats, but this only supports one, so far.

Made a requirements.txt and a .gitignore
main.py runs analyzer.run()